Final tweaks for the refactor
