+++ src/components/SwipeCard.js
@@ -1,41 +1,41 @@
-import React, { Component } from 'react';
-import SwipeCardMotion from './SwipeCardMotion';
-import { motion } from "framer-motion";
+import React from 'react';
 
-class SwipeCard extends Component {
-    constructor(props) {
-        super(props);
-        this.swipeClassifier = this.swipeClassifier.bind(this);
-        console.log(this.props);
-    }
+import { motion, useMotionValue, useTransform } from "framer-motion";
 
-    swipeClassifier(e, i) {
+export const SwipeCard = (props) => {
+
+    function swipeClassifier(e, i) {
         if (i.offset.x >= 250) {
             console.log(false);
-            this.props.unmountCard();
+            props.unmountCard();
         } else if (i.offset.x <= -250) {
             console.log(true);
-            this.props.unmountCard();
+            props.unmountCard();
         }
     }
 
-    render() {
+    const x = useMotionValue(0);
+    const xInput = [-200, 0, 200];
+    const backgroundColor = useTransform(x, xInput, [
+        "#FFDCDC",
+        "#FFFFFF",
+        "#E6F4DC"
+    ]);
+    const rotate = useTransform(x, xInput, [-45, 0, 45]);
 
-        return (
-            <SwipeCardMotion>
-                <motion.div
-                    // drag
-                    // dragConstraints={{ left: 0, right: 0, top: 0, bottom: 0 }}
-                    className="swipe-card"
-                    onDrag={(e, i) => {
-                        this.swipeClassifier(e, i);
-                    }}
-                >
-                    test 2
-                </motion.div>
-            </SwipeCardMotion>
-        );
-    }
+    return (
+        <motion.div
+            drag
+            dragConstraints={{ left: 0, right: 0, top: 0, bottom: 0 }}
+            onDrag={(e, i) => {
+                swipeClassifier(e, i);
+            }}
+            className="swipe-card"
+            style={{ x, backgroundColor, rotate }}
+        >
+            {props.children}
+        </motion.div>
+    )
 }
 
 export default SwipeCard;
